feat(data/finset): sup_eq_bind
`finset.sup s f` is equal to `finset.bind s f` when `f : α → finset β`
is an indexed family of finite sets.  This is a proof of that with
a couple supporting lemmas.  (There might be a more direct proof
through the definitions of `sup` and `bind`, which are eventually in
terms of `multiset.foldr`.)