feat(tactic/norm_swap): simplify numeral swaps (#5637)
Explicitly applied swap equalities over `nat` can be proven to be true using `dec_trivial`. However, if occurring in the middle of a larger expression, a separate specialized hypothesis would be necessary to reduce them down. This is an initial attempt at a `norm_num` extension which seeks to reduce down expressions of the `swap X Y Z` form. Handles `nat`, `int`, `rat`, and `fin` swaps.



Co-authored-by: Yakov Pechersky <pechersky@users.noreply.github.com>