feat(measure_theory/integration): add theorems about the product of independent random variables (#6106)
Consider the integral of the product of two random variables. Two random variables are independent if the preimage of all measurable sets are independent variables. Alternatively, if there is a pair of independent measurable spaces (as sigma algebras are independent), then two random variables are independent if they are measurable with respect to them.




Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>