Experiment with Eric's "open the floodgates" suggestion
Despite initial reluctance, I am beginning to be persuaded. The
names are not serious suggestions, especially as if this works, what
we are really defining is non-unital algebras since we can make the
existing `algebra` definition allow for non-associativity for free.

For now I just want to see whether this compiles.