refactor(algebra/big_operators, *): specialize `finset.prod_bij` to `fintype.prod_equiv`
Often we want to apply `finset.prod_bij` to the case where the product is taken
over `finset.univ` and the bijection is already a bundled `equiv`. This PR adds
`fintype.prod_equiv` as a specialization for exactly these cases, filling in
most of the arguments already.

I still want to go through the usages of `prod_bij` to see where the new
`prod_equiv` can be usefully applied.