feat(analysis/normed_space/banach): a range with closed complement is itself closed (#6972)
If the range of a continuous linear map has a closed complement, then it is itself closed. For instance, the range can not be a dense hyperplane. We prove it when, additionally, the map has trivial kernel. The general case will follow from this particular case once we have quotients of normed spaces, which are currently only in Lean Liquid.

Along the way, we fill several small gaps in the API of continuous linear maps.