feat(algebra/polynomial): generalize to `multiset` products
This PR generalizes the results in `algebra.polynomial.big_operators` to
sums and products of multisets.

In order to easily derive the equivalent `finset` results, the new lemmas are
stated in terms of `(t.map f).prod`, rather than `t.prod` directly. However,
there was already a `nat_degree_multiset_prod` that directly uses `t.prod`,
so now it's a bit inconsistent. I guess the alternative would be to give the
lemmas in both forms. (Called `multiset_map_prod` and `multiset_prod`
respectively?)